In this version, I added the "/q" flag to the del command,
which makes the delete command run quietly without displaying a confirmation prompt. I added the "@echo off" command at the start of the script to hide the commands being executed in the command prompt window. I also added the "echo" command to confirm that the files have been deleted successfully, and "pause" command at the end of the script to keep the command window open so that the user can see the confirmation message.